create package for  AllocateStrategy and modify the doc according to comments
