+++ core/src/test/java/tachyon/worker/hierarchy/AllocateStrategyTest.java
@@ -25,6 +25,9 @@ import org.junit.Test;
 import tachyon.TestUtils;
 import tachyon.UnderFileSystem;
 import tachyon.worker.BlockHandler;
+import tachyon.worker.allocation.AllocateStrategies;
+import tachyon.worker.allocation.AllocateStrategy;
+import tachyon.worker.allocation.AllocateStrategyType;
 import tachyon.master.BlockInfo;
 import tachyon.util.CommonUtils;
